requirements.txt
# requirements.txt
flask
python-dotenv
requests
pytesseract
opencv-python
Pillow
unstructured-client
fitz
